pyproject.toml
[project]
name = "sharpa"
version = "0.1.0"
description = "Reasoning and planning layer: goal frame synthesis, verification, and exploration"
authors = [
    { name = "para-lost", email = "18801102729@163.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "structlog>=24.0",
    "pillow>=10.0",
    "pydantic-settings>=2.13.1",
    "openai>=2.24.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]
sim = [
    "gymnasium==0.29.1",
    "mani_skill>=3.0.0b",
    "stable-baselines3>=2.3",
    "torch>=2.3",
    "setuptools>=69",
]

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.2",
]